Move packet model data and modifiers
